Turning on unit tests build and run during precommit.

+++ tests/unit/common.h
@@ -16,13 +16,12 @@
 #ifndef COMMON_H
 #define COMMON_H
 
-#include "jerry-libc.h"
 #include "literal.h"
 
 #define NAME_TO_ID(op) (__op__idx_##op)
 
 #define __OPCODE_SIZE(name, arg1, arg2, arg3) \
-  sizeof (__op_##name) + 1,
+  (uint8_t) (sizeof (__op_##name) + 1),
 
 #define LP(s) create_literal_from_str_compute_len (s)
 #define NUM(s) create_literal_from_num (s)
